refactor: 공통적으로 사용되는 Article 관련 컴포넌트는 components/Common/ 내부로 이동
