+++ src/components/Common/Article/ArticleImgSlider/index.js
@@ -1,0 +1,3 @@
+import ArticleImgSlider from "components/Common/Article/ArticleImgSlider/ArticleImgSlider";
+
+export default ArticleImgSlider;
