+++ src/components/Common/Article/ArticleMain.tsx
@@ -1,7 +1,7 @@
 import PopHeart from "components/Common/PopHeart";
 import React, { useMemo, useState } from "react";
 import styled from "styled-components";
-import Username from "../../Common/Username";
+import Username from "../Username";
 import { useAppDispatch, useAppSelector } from "app/store/Hooks";
 import { modalActions } from "app/store/ducks/modal/modalSlice";
 import { getMiniProfile } from "app/store/ducks/modal/modalThunk";
