--- src/components/Home/Article/ArticleImgSlider/index.js
@@ -1,3 +1,0 @@
-import ArticleImgSlider from "components/Home/Article/ArticleImgSlider/ArticleImgSlider";
-
-export default ArticleImgSlider;
