+++ src/components/Home/HomeSection.tsx
@@ -1,8 +1,8 @@
 import { getHomeArticles } from "app/store/ducks/home/homThunk";
 import { useAppDispatch, useAppSelector } from "app/store/Hooks";
+import Article from "components/Common/Article";
 import ExtraLoadingCircle from "components/Home/ExtraLoadingCircle";
 import { useEffect } from "react";
-import Article from "./Article";
 
 const HomeSection = () => {
     const { articles, isLoading, isExtraArticleLoading } = useAppSelector(
